+++ core/MapNode.cpp
@@ -22,7 +22,7 @@
 namespace warmonger {
 namespace core {
 
-MapNode::MapNode(QObject* parent, long id)
+MapNode::MapNode(QObject* parent, int id)
     : WObject(parent, id)
 {
 }
